implement filtering for uast
