+++ frontend/src/components/CodeViewer.js
@@ -1,13 +1,10 @@
 import React, { Component } from 'react';
-import { Modal, Button } from 'react-bootstrap';
+import { Modal } from 'react-bootstrap';
 import PropTypes from 'prop-types';
 import SplitPane from 'react-split-pane';
-import UASTViewer, {
-  Editor,
-  withUASTEditor,
-  languageToMode
-} from 'uast-viewer';
+import { Editor, withUASTEditor, languageToMode } from 'uast-viewer';
 import Switch from 'react-switch';
+import UASTViewerPane from './UASTViewerPane';
 import api from '../api';
 import './CodeViewer.less';
 import CloseIcon from '../icons/close-query-tab.svg';
@@ -45,131 +42,6 @@ EditorPane.propTypes = {
   language: PropTypes.string,
   handleLangChange: PropTypes.func.isRequired,
   editorProps: PropTypes.object
-};
-
-const ROOT_ID = 1;
-const SEARCH_RESULTS_TYPE = 'Search results';
-
-function getSearchResults(uast) {
-  if (!uast) {
-    return null;
-  }
-
-  const rootNode = uast[ROOT_ID];
-  if (!rootNode) {
-    return null;
-  }
-
-  if (rootNode.InternalType === SEARCH_RESULTS_TYPE) {
-    return rootNode.Children;
-  }
-
-  return null;
-}
-
-function NotFound() {
-  return <div>Nothing found</div>;
-}
-
-function UASTViewerPane({
-  uastViewerProps,
-  showLocations,
-  useCustomServer,
-  customServer,
-  filter,
-  handleShowLocationsChange,
-  handleUseCustomServerChange,
-  handleCustomServerChange,
-  handleFilterChange,
-  handleSearch
-}) {
-  const searchResults = getSearchResults(uastViewerProps.uast);
-  const rootIds = searchResults || [ROOT_ID];
-
-  let content = null;
-  if (uastViewerProps.uast) {
-    if (searchResults && !searchResults.length) {
-      content = <NotFound />;
-    } else {
-      content = (
-        <UASTViewer
-          {...uastViewerProps}
-          rootIds={rootIds}
-          showLocations={showLocations}
-        />
-      );
-    }
-  }
-
-  return (
-    <div className="uast-viewer-pane">
-      <div className="show-locations-wrapper">
-        <label>
-          <input
-            type="checkbox"
-            checked={showLocations}
-            onChange={handleShowLocationsChange}
-          />
-          <span>Show locations</span>
-        </label>
-        <label>
-          <input
-            type="checkbox"
-            value={useCustomServer}
-            onChange={handleUseCustomServerChange}
-          />
-          <span>Custom bblfsh server</span>
-        </label>
-        {useCustomServer ? (
-          <input
-            type="text"
-            value={customServer}
-            onChange={handleCustomServerChange}
-          />
-        ) : null}
-      </div>
-      <div className="uast-query-wrapper">
-        <form
-          onSubmit={e => {
-            e.preventDefault();
-            handleSearch();
-          }}
-        >
-          <input
-            type="text"
-            placeholder="UAST Query"
-            value={filter}
-            onChange={handleFilterChange}
-          />
-          <Button bsStyle="gbpl-secondary" type="submit">
-            SEARCH
-          </Button>
-          <Button
-            className="edit-query"
-            bsStyle="gbpl-primary-tint-2-link"
-            href="https://doc.bblf.sh/using-babelfish/uast-querying.html"
-            target="_blank"
-          >
-            HELP
-          </Button>
-        </form>
-      </div>
-      {content}
-    </div>
-  );
-}
-
-UASTViewerPane.propTypes = {
-  uastViewerProps: PropTypes.object,
-  showLocations: PropTypes.bool,
-  useCustomServer: PropTypes.bool,
-  customServer: PropTypes.string,
-  filter: PropTypes.string,
-  handleShowLocationsChange: PropTypes.func.isRequired,
-  handleUseCustomServerChange: PropTypes.func.isRequired,
-  handleCustomServerChange: PropTypes.func.isRequired,
-  handleFilterChange: PropTypes.func.isRequired,
-  handleSearch: PropTypes.func.isRequired
 };
 
 function EditorUASTSpitPane({
