+++ frontend/src/components/UASTViewerPane.less
@@ -1,0 +1,67 @@
+.uast-viewer-pane {
+    display: flex;
+    flex-direction: column;
+    height: 100%;
+
+    .show-locations-wrapper {
+        flex: 0 0 auto;
+
+        min-height: 45px;
+        padding: 5px 0px 5px 20px;
+        background: @primary-tint-4;
+        border-bottom: solid 1px fade(@primary-tint-2, 50%);
+
+        font-size: 14px;
+
+        label {
+            font-weight: normal;
+            margin-right: 40px;
+
+            span {
+                margin-left: 10px;
+            }
+
+            &:last-of-type {
+                margin-right: 10px;
+            }
+        }
+
+        input {
+            .blue-header-element;
+
+            font-family: @monospace-font;
+        }
+    }
+
+    .uast-query-wrapper {
+        flex: 0 0 auto;
+
+        background: @primary-tint-4;
+        padding: 10px 15px 10px 20px;
+        overflow: hidden;
+
+        form {
+            display: flex;
+
+            input {
+                .blue-header-element;
+
+                flex: 1 1 auto;
+                min-width: 25px;
+
+                color: @primary-tint-1;
+                font-family: @monospace-font;
+                font-size: 14px;
+                font-weight: bold;
+            }
+
+            button {
+                margin-left: 15px;
+            }
+        }
+    }
+
+    .uast-viewer {
+        flex: 1 1 auto;
+    }
+}
